Try to more consistently use `size_t` for argument of `nclistlength`

+++ libdap2/cdf.c
@@ -71,7 +71,7 @@ computecdfnodesets(NCDAPCOMMON* nccomm, CDFtree* tree)
 NCerror
 computevarnodes(NCDAPCOMMON* nccomm, NClist* allnodes, NClist* varnodes)
 {
-    unsigned int i,len;
+    size_t i, len;
     NClist* allvarnodes = nclistnew();
     for(i=0;i<nclistlength(allnodes);i++) {
 	CDFnode* node = (CDFnode*)nclistget(allnodes,i);
